+++ src/migrations/20220530233030-create-plant.js
@@ -12,22 +12,22 @@ module.exports = {
                 type: Sequelize.STRING,
             },
             min_ph: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.FLOAT,
             },
             max_ph: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.FLOAT,
             },
             min_tds: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.FLOAT,
             },
             max_tds: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.FLOAT,
             },
             min_ec: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.FLOAT,
             },
             max_ec: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.FLOAT,
             },
             createdAt: {
                 allowNull: false,
